refactor: Pass input items as children to the input component to prevent extra re-renders when error state changes
