+++ src/features/spy/command.tsx
@@ -4,6 +4,7 @@ import FieldLabel from '@/components/field-label';
 import Input from './input';
 import { MODES } from '@/system/spy';
 import { useModeStore } from '@/system/store';
+import InputItems from './input-items';
 
 function Command() {
   const currentMode = useModeStore();
@@ -15,7 +16,9 @@ function Command() {
           return mode.fields.map((field, i) => {
             return (
               <FieldLabel key={i} title={field.name} hint={field.hint}>
-                <Input name={field.name} type={field.type} placeholder={field.placeholder} />
+                <Input name={field.name} type={field.type} placeholder={field.placeholder}>
+                  <InputItems name={field.name} />
+                </Input>
               </FieldLabel>
             );
           });
